+++ ripozo/viewsets/relationships/relationship.py
@@ -78,7 +78,8 @@ class Relationship(object):
         related_properties = self._map_pks(properties)
         resource = None
         if related_properties:
-            resource = self.relation(properties=related_properties, query_args=query_args)
+            resource = self.relation(properties=related_properties, query_args=query_args,
+                                     include_relationships=self.embedded)
         if self.required and (not resource or not resource.has_all_pks):
             raise RestException('The relationship {0} could not construct a valid {1}'
                                 ' with all of its pks.  Properties'